chore: add link checking in CI
This adds a link-checking action run on a weekly basis using lychee:

The check can also be triggered manually using the Github CLI or the 'Actions' tab interface for contributors of the repo.

This does not act on PRs but rather publishes / updates an issue on the repo with the link check report.