fix: remove panic in process_headers, test CertificateWaiter more thoroughly
This introduces a long test firing up the `CertificateWaiter` and all dependent components,
and feeding it valid certificates in inverse (decreasing) round order, simulating catching up from a late node.

It tweaks a few mock functions to generate those certificates and their headers in the first place.

Along the way, this discovers a panic in this catch-up path:
- process_certificate calls process_header,
- process_header can return with `Ok(())` in case the node has missing parents / payload,
- but in the case it doesn't, we will try to vote for the header,
- that vote may be a duplicate vote if
    + the node was among thevoters for this certificate,
    + a restart has blown the node's in-memory `last_voted` table,
- we change this panic to instead generate the following message at the error level:
```
2022-07-13T12:36:25.607333Z ERROR process_header: primary::core: Failed to process our own vote: Authority vq2gYSbHjZi0oaafbuYYlpTw9HUVONqCTxrcixShtWI= appears in quorum more than once
```