Increase base consensus timeout and make it a config (#4532)
Sui has a timeout on sending transaction to the nw consensus and automatically adjusts the timeout based on recent latency from nw it has seen in case of success, or linearly increases in case of timeout.

In the stress test network we see quite a few of nw timeouts from the sui transaction submission side, and we indeed [see](http://grafana.shared.internal.sui.io:3000/goto/wqnUWbGVz?orgId=1) that control delay is adjusted accordingly to the timeouts and nw is not stuck.

However, it seems that current timeout adjustment logic drops timeout pretty quickly which causes timeouts again.

While this is not end of the world, it is noise and probably suboptimal.

It also can be noticed that current base timeout value is very close to the actual consensus delay, so the timeout has to be frequently adjusted. Therefore, a short-medium term fix (this PR) is just to bump the delay step for timeout calculations so that it is further away from median nw latency.

In the future we can improve timeout calculation by using weighted avg over a longer window instead of weighting current and next timing only.