[bullshark]: do not insert certificates past committed round (#5238)
This check ensures that certificates past committed round for the given origin are not inserted in the consensus DAG.

This is believed to be a source of consensus fork we observe in the private testnet:

(1) Primary is assumed to be able to send repeating certificates to ConsensusProtocol::process_certificate
(2) Currently ConsensusProtocol::process_certificate inserts certificate to hashmap unconditionally
(3) When commit is processed, certificates with round *less than* last committed round are removed in ConsensusState::update
(4) When dag is ordered, certificates with round **equal to** last committed round are ignored during sequencing

The current sequence opens loophole when certificate to round **previous to** last committed is submitted into consensus, it might be sequenced again later. This is in line with current logs we observe in testnet.

This PR enforces invariant that dag does not contain nodes with round **less than** last committed for the given node.