[sui core] check for effects before verifying cert
I believe this will be needed after we have reconfig. The signature checking logic looks at the *current* committee, but this will make signature checking fail if you're looking at a cert w/ a different committee from a previous epoch.

In addition, this saves a signature check on the happy path when someone is asking for effects from a cert we've previously executed.