[narwhal] fetch missing certificates based on GC round and existing certificates (#5522)
Currently `CertificateWaiter` requests certificates based on last committed rounds. But when processing certificates, all parents of a certificate are still required. This is incorrect, as outlined [here](https://github.com/MystenLabs/sui/pull/5058#discussion_r999297817). Basically a certificate may not be needed by the DAG because it is below the last committed round of its authority, but it is still needed as a parent to another certificate.

As a fix, fetch certificates based on current GC round and existing certificates above GC round. Also, restore the previous logic that a certificate is considered to have parents if its parents are at the GC round threshold.