[authority] introduce notify_read for effects (#5913)
This commit introduces `EffectsNotifyRead` trait that has `notify_read(Vec<TransactionDigest>) -> SuiResult<Vec<TransactionEffects>>`.

This trait is implemented on the `AuthorityState`: we register notification channels first, then inspect if effects are stored in the database. We cancel notifications for all the effects found in the database and wait for others.

When effects are written in the database, we notify all pending registrations.

Registering notifications before probing the database this sequence is necessary to avoid race conditions between putting registration and writing effects to database.

This PR tries to have most of notify-read code separated into `AuthorityNotifyRead` struct and only change to `authority.rs` itself is to call `AuthorityNotifyRead::notify` when effects are persisted in the database.

Some design notes:

This is implemented as specialized code rather than some code in the `DBMap`. While it is possible to move this to `DBMap` level it is not super trivial given that effects are written as part of the batch. I think we can still consider moving this to DBMap when mysten-infra is merged to sui and *if* there are some other plausible use cases for this outside of effects store. But I prefer to merge this as is for now, without doing too much premature generalization.