[consensus] Use SuiTxValidator to verify signatures (#6626)
This moves signature verification for consensus messages to narwhal payload verification step, effectively removing signature verification from the consensus handler critical path.

To elaborate, it is safe to skip signature verification in consensus handler, because in order to create certificate in narwhal, header needs to be signed by 2f+1 authorities, which only sign header if they have all batches stored locally. Additionally, authorities only store batches locally if all signatures in the batch are correct according to configured `TxVerifier`.

Additional benefit of this approach is that because batches are processed in parallel, this gives parallel verification of signatures for free.

Note: one hesitation for using `SuiTxValidator` initially was that it does not have any mechanism for signature verification deduplication. However, after #6588 is merged we have nearly no amplification for consensus messages, and it is ok to just use `SuiTxValidator` as is.