[reconfig] Restart authority store using perpetual epoch (#6938)
We cannot restart the authority store using the epoch/committee from
on-chain state, because that could get ahead of the actual current epoch
of the node, which should only change at reconfiguration.
This PR adds a persisted current epoch to the perpetual tables, and use
that as source of truth. It's only updated at reconfiguration. And when
we restart, we rely on it to locate the proper epoch table.