[reconfig] Add epoch id to ConsensusHandler (#6956)
`ConsensusHandler` is the interface where Narwhal talks to Sui. To
ensure strong data consistency, for instance, when creating checkpoints
we want to know exactly which epoch we are creating it for, we need a
top-down passed epoch id that doesn't change in the middle.
This PR adds epoch id to ConsensusHandler. ConsensusHandler is
reconstructed each time we reconfigure, to make sure it's always
associated with a single constant epoch. In order to do so,
ConsensusHandler needs to be moved out of Narwhal config, and put into
the message we send to the narwhal manager.