[reconfig] Pass epoch id from ConsensusHandler to all transaction path (#6962)
This PR uses the epoch id in ConsensusHandler and pass it down to all
paths that processes the transaction.
A separate PR will pass it for all checkpoint related functions.
This PR also moves a few functions into the per epoch table since it's
mostly operating there. Also added a new epoch table load function that
takes an expected epoch id and checks against it. Eventually we should
replace all uses with it.