[reconfiguration] Avoid race condition between execution and reconfiguration. (#7054)
This PR adds a separate RwLock to authority state to guard execution
epoch. See comments on `reconfigure()` and `process_certificate()`
methods on the order of this epoch change and epoch on the epoch store.

This ensures that before we attempt to revert transactions, no
transaction is being executed in parallel.

Also, after epoch change, if any transaction were scheduled for the
previous epoch, it will be rejected as execution epoch carries the epoch
number and `process_certificates` checks that epoch number.