Distinguish insufficient gas errors (#6981)
Before this PR, `InsufficientGas` is an enum in both
`ExecutionFailureStatus` and `SuiError` but the meanings are different:
- `SuiError::InsufficientGas` means the balance of gas object is lower
than the specified gas budget
- `ExecutionFailureStatus::InsufficientGas` means that the txn failed
b/c the actual gas cost has gone beyond `gas_budget`

As a result, sometimes it was a bit tricky to tell the exact error on
the client side if the type info of `ExecutionFailureStatus` or
`SuiError` was missing. This PR renames the one in `SuiError` to
`InsufficientGasForGasBudget` to distinguish them.