Fix reconfig_with_revert_end_to_end_test to always use the longest-waiting validator as the revert target (#7182)
This improves a basic source of flakiness in the test - this ensures (or
at least makes it very likely) that the to-be-reverted tx in the test is
not submitted to consensus before the end-of-epoch messages are sent by
the other 3 validators. Previously, the order was a matter of random
chance, so the test would fail if we happened to pick the most eager
validator to be the reverter.