[reconfiguration] Do not revert transactions executed via state sync (#7255)
There is a condition where we can attempt to revert final transaction:
we can have transaction that we proposed to consensus, but epoch ended
via state sync before we reached end of consensus stream. In that case
we might attempt to revert transaction that is actually final.

This PR fixes this problem by recording all executed transactions in the
epoch table and then verifying that only transactions that were not
executed as part of checkpoint process are reverted.

This diff works under assumption that all checkpoints(both for
checkpoints that are locally created and synchronized from remote) go
through checkpoint executor.