Ensure deletes invoked while pruning object table are non overlapping (#7387)
Today when we invoke range delete on the objects table, we trigger
deletes with overlapping ranges which causes RocksDB to fragment them to
during read. For example, if at time T1 object table is laid out with
two keys A and B both with 5 versions each like [(A, 0), (A, 1), (A, 2),
(A, 3), (A, 4), (B, 0), (B, 1), (B, 3), (B, 4), (B, 5)] - we submit
range deletes for ranges [(A, 0), (A,2)] and [(B, 0), (B,2)]. And later
at time T2, if both gained newer 3 versions each such that table is laid
out as [(A, 3), (A, 4), (A, 5), (A, 6), (A, 7), (B, 3), (B, 4), (B, 5),
(B, 6), (B, 7)] - we submit range deletes with ranges [(A, 0), (A,5)]
and [(B, 0), (B,5)]. If there are multiple such pruning runs before
compaction could get to cleaning up, we end up with lots of overlapping
ranges i.e. [(A, 0), (A, 2)] is overlapping with [(A, 0), (A, 5)], [(B,
0), (B, 2)] is overlapping with [(B, 0), (B, 5)], etc which causes it to
do a lot of range fragmentation (as we saw in the stack trace as well).
A simple fix is to not do overlapping range deletes. In the above
example, after doing a delete of [(A, 0), (A, 2)] next range delete will
be for [(A, 3), (A, 5)] and so on. This way delete ranges will never
have to overlap and prevent RocksDB from doing extra work during get.