[1/x][move] Expose Clock to Move code (#7922)
Expose a shared object to read time from -- `Clock` -- that is created
during genesis, and an accessor for move. This commit does not introduce
the system transaction that updates the timestamp held in this object,
so its timestamp will remain at `0`, except in tests which are free to
increment it.

Also introduces a sui-specific verification rule that entry functions
may only accept a `Clock` by immutable reference (not by value or
mutable reference). This is to complement RWLock work:

- RWLock eases contention on shared objects that are read much more
often than they are written, but it does not prevent people from taking
a mutable reference to the shared object even if they only need to read
it.

- Only the system transaction legitimately needs to take the `Clock` by
mutable reference, so the new verification rule prevents the creation of
entry functions that could cause unnecessary contention on the `Clock`,
intentionally or otherwise.

- Validators will also prevent the signing of transactions that accept
the `Clock` by mutable reference (excluding the system transaction to
update it), as a further precaution -- not part of this commit.

## Breaking Changes

- New Sui Framework Module `sui::clock`,
- New genesis object `0x6: sui::clock::Clock` (requires re-genesis).

## Test Plan

New verifier and genesis snapshot tests:

```
sui$ cargo nextest run
```