[core] Make certificate optional in validator handle tx/cert API (#8359)
With finalized transactions, validator can no longer guarantee to return
transaction certificates. Hence we need to change validator grpc APIs to
return optional certificate in the executed case.
This PR makes such change. The core of the change is in messages.rs.
For handle_transaction, this PR also makes an optimization where we
don't actually need to return the transaction data, but only the
transaction signature in the signed and cert case.