[Narwhal] move `Certificate` processing from `Core` to `Synchronizer` (#8270)
Currently much of `Certificate` processing logic lives in `Core`, which
has the following drawbacks:
- Processing is single threaded, even if much of `Certificate`
processing can be parallelized, e.g. verifications, checking parents,
etc.
- It is hard to evolve, e.g. if we want to lookup suspended certificates
while checking existence of parents, a new channel need to be added.

This PR moves the `Certificate` processing logic into `Synchronizer`,
which is now responsible for synchronizing DAG, not just the missing
parent certificates.

In future, additional data structures will be added to `Synchronizer` to
keep track of suspended certificates.

Majority of the changes are the required plumbing. The most significant
change is that certificate broadcast is re-implemented.

#8288